pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visionaid"
version = "0.1.0"
description = "LLM vision aid auditing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic~=0.84.0",
    "beautifulsoup4>=4.9.2",
    "lxml>=4.9.0",
    "openai>=1.0.0",
    "python-dotenv~=1.2.1",
    "requests>=2.32.2",
]

[tool.setuptools]
packages = ["vision_aid"]